Start adding third party files but properly this time
