+++ csfieldguide/static/interactives/box-translation/package.json
@@ -1,0 +1,9 @@
+{
+    "name": "csfg-interactive-box-translation",
+    "version": "1.0.0",
+    "private": true,
+    "dependencies": {
+        "three": "0.101.1",
+        "tweenjs": "1.0.2"
+    }
+}
