+++ csfieldguide/templates/interactives/box-translation.html
@@ -51,20 +51,6 @@
         </div>
       </div>
     </div>
-
-    <!-- threejs -->
-    <script src="../base-files/js/third-party/threejs/three.min.js"></script>
-
-    <!-- threejs libraries -->
-    <script src="../base-files/js/third-party/threejs/Detector.js"></script>
-    <script src="../base-files/js/third-party/threejs/Projector.js"></script>
-    <script src="../base-files/js/third-party/threejs/OBJLoader.js"></script>
-
-    <script src="../base-files/js/third-party/tween.js"></script>
-
-    <!-- custom code for interactive -->
-    <script src="js/translation-script.js"></script>
-    <script src="../base-files/js/translation-rotation-interactives/script.js"></script>
   </div>
 {% endblock html %}
 
@@ -73,5 +59,5 @@
 {% endblock css %}
 
 {% block js %}
-  <script type="text/javascript" src="{% static 'interactives/box-translation/js/box-translation.js' %}"></script>
+  <script type="text/javascript" src="{% static 'js/translation-rotation-interactives/script.js' %}"></script>
 {% endblock js %}
